Safari-611.1.10/LayoutTests/fast/css/case-transform.html

<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
<script>if (window.testRunner) testRunner.dumpAsText();</script>
</head>
<body>
<p>lower(Ağ€) = <span style="text-transform:lowercase">Ağ€</span> (should be ağ¨)</p>
<p>upper(ağ¨) = <span style="text-transform:uppercase">ağ¨</span> (should be Ağ€)</p>
<p>lower(ÎšÎŸÎ£ÎœÎŸÎ£) = <span style="text-transform:lowercase">ÎšÎŸÎ£ÎœÎŸÎ£</span> (should be ÎºÎ¿ÏƒÎ¼Î¿Ï‚)</p>
<p>upper(ÃŸ) = <span style="text-transform:uppercase">ÃŸ</span> (should be SS)</p>
<p>upper(Å‰) = <span style="text-transform:uppercase">Å‰</span> (should be Ê¼N)</p>
<p>upper(Ç°) = <span style="text-transform:uppercase">Ç°</span> (should be JÌŒ)</p>
<p>upper(ï¬ƒ) = <span style="text-transform:uppercase">ï¬ƒ</span> (should be FFI)</p>
<p>lower(Ä²) = <span style="text-transform:lowercase">Ä²</span> (should be ij)</p>
</body>
</html>
