third_party/blink/web_tests/platform/generic/fast/css/case-transform-expected.txt

lower(Ağ€) = ağ¨ (should be ağ¨)

upper(ağ¨) = Ağ€ (should be Ağ€)

lower(ÎšÎŸÎ£ÎœÎŸÎ£) = ÎºÎ¿ÏƒÎ¼Î¿Ï‚ (should be ÎºÎ¿ÏƒÎ¼Î¿Ï‚)

upper(ÃŸ) = SS (should be SS)

upper(Å‰) = Ê¼N (should be Ê¼N)

upper(Ç°) = JÌŒ (should be JÌŒ)

upper(ï¬ƒ) = FFI (should be FFI)

lower(Ä²) = Ä³ (should be ij)
